requirements.txt
Flask==2.0.1           # For web application framework
Werkzeug==2.0.1        # For utilities including secure filename handling
opencv-python==4.11.0.86  # For image processing capabilities
numpy>=1.21.2         # For numerical operations
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Pillow==10.3.0  # For image annotation and processing